fix dock audio enable dialog
The dock audio enable setting is not saved if the user
dismisses the dialog without checking or unchecking the check box.
The dialog is then displayed each time the device is docked until the
user checks/unchecks the dialog or goes to sound settings to enable
or disable dock audio.

Bug 7518165.

Change-Id: Ifbc6179e78b64b5b7577ac750acf9846525e47c5