Replacing accessibility service permissions with capability attributes.
Accessibility services can perform special operations such as retrieve
the screen content, enable explore by touch, etc. To ensure the user
is aware that the service will perform special operations we were using
permissions. However, the special operations cannot be performed unless
the service is really enabled by the user and it is at this point that
we want to notify the user about the service capabilities.

This change adds the list of the service capabilities to the enable
warining dialog such that the user can asses whether to he/she is fine
with this service performing certain special operations.

Main change:https://googleplex-android-review.googlesource.com/#/c/298110/

bug:8633951

Change-Id: I1617fa82126b9a674cb74a15d7004a7cc007c7d3