Fix profile picture picking for restricted profiles
Due to the new file picker UI and incorrect assumptions of access to sd card,
have to use content provider Uris instead of file paths.

Also makes the cropping robust in the event of not having a cropping intent
on the platform.

This should make the code play well with third party image
and camera apps as well.

Bug: 10666584

Change-Id: Ie8d834fe7aac96bc14829a7be084512a15ef4001