Make printers for a print service not react to press.
We show a list of printers for the a print service in the
print service settings activity. This list is there to
inform the user of which printers are available and
optionally offers adding a printer affordance. The printer
list items do nothing so they should not react to press
and create a false impression they do.

bug:10983508

Change-Id: I43b19ebb0acf6662c489826c192fd6daf8ad103d