Wi-Fi Direct setting UI: handle p2p disable state
This is not a normal use case as we can't enter Wi-Fi Direct setting when p2p
is disabled. However we can make this happen by entering Wi-Fi Direct setting
screen then disabling wifi via quick settings.

Previous UI implementation didn't handle p2p enable->disable change completely,
leaving the screen out-of-date until we change the screen orientation.

Now we always attach mThisDevicePref/mPeersGroup/mPersistentGroup to the screen,
but just disable them when p2p is disabled. Another advantage of doing this is
to avoid b/11050625, which occurs when a PreferenceGroup is not attached to the
PreferenceScreen.

Bug: 11133069
Change-Id: I9be7b6e396d7677b5a33df8e6294b4a0aa6bd389