Check for presence of carrier app in Mobile Plan.
When the user chooses Mobile Plan in Settings, first do a check to see if
a valid carrier app for the current MCC/MNC exists. If so, launch that app
instead of checking for a provisioning URL or hoping for a DNS redirect in
the browser. If no such app exists, continue on as normal.

Bug: 9623212
Change-Id: I81538063d4cdd147948a83ffc0c2641174ba9001