Remove duplicates into Search results
In schema version 100, we were using data_title and data_summary for
storing a normalized version of the title and summary.

This was leading to have duplicate results.

Now (in schema version 101), we are introducing two new colums for
storing normalized title and summary. We are also doing a new MATCH
query on those two new columns and that fixes the duplicates issue.

Change-Id: I609675cdd1a47a1a29595f0fe8b87ad96c6e4522