Improve power fail/reset
On power fail/reset, an encrypted device will sit at the enter password screen
indefinitely, chirping. This is designed to attract the attention of the user.

However, it also flattens the battery, and the user who's attention is not
drawn will discover a discharged phone the next morning. We have had many
complaints about this.

Keep current functionality, but power down after 10 minutes. It's a compromise,
but seems reasonable.

@bug 12582489

Change-Id: I895c0147bed978ecf6984af2c748f971dfa0d221