Enable OEM unlock checkbox in Developer Settings
For Volantis+ devices, we will give users
the ability to enable OEM unlock through
Developer Settings. To do so, we must write
the value to the last byte of a special partition
that does not get erased even after factory reset.

This feature will only be available on devices
with the persistent data partition, thus the checkbox
is only visible for devices that meet this requirement.

This depends on
https://googleplex-android-review.git.corp.google.com/#/c/495350/

DD: go/factory-reset

Bug: 14288780
Change-Id: I8f80b950bc101a1067912faf221391bf0dd826b7