Link WifiAssistant card with actual settings.
Only show the card if we haven't already enabled an assistant, and if
there's one available on the device to enable. (For now, we enable the
first one; in the future, we can offer a selection between
assistants).

When the user opts in, we either show the standard system activity or
the custom opt-in activity if the assistant has defined one. In either
case, we only dismiss the card if the child activity indicates that
opt-in was successful.

Bug: 16376277
Change-Id: Ic92c7676c563ab8c8fb695117a053004f15f7d25