Allow CryptKeeper dialogs to rotate on tablets
Tablets may be used in docking stations thus need to support landscape mode.
Phones will not be, and having them rotate is both ugly and annoying.
Also this matches lock screen behavior.

The attribute screenOrientation is not themeable, so the only way of doing
this is programmatic.

Test for truth of bool which is only set for sw600dp devices and
clear the nosensor setting from the manifest if set.

Bug: 16405551
Change-Id: I5b3f945b89402e5aa39642c2c06e98f9ca1ce298