Bring screen lock setting back to screen pinning
The material design of the request has removed this option from
the dialog, and now needs to be in settings.  When screen
pinning is off it shows a description of how to use the feature,
however when on it shows a switch to control whether it locks
the device on exit.  This setting defaults to on if there is a
lock pin/pattern.

Bug: 16957435
Change-Id: I59e0d64964cee4302e6ecff59e622cfe11422807