Start activity for the correct user in Settings > Data Usage
The 'app settings' button on the app details view now starts the app settings activity
for the correct user, i.e. the managed profile's app settings if the list item corresponds
to the managed profile's app. Formerly it always incorrectly started up the settings screen
for the primary user's app.

Bug:16727726
Change-Id: Ib31d45f4201fb26b5231e4e57c49858e7e41efc9