Several fixes to new EncryptionInterstitial dialog.
The code now observes whether accessibility is turned on when
deciding the default state.

Additionally, it fixes a bug where the user can back out of
EncryptionInterstitial and leave the setting in a bad state.
We now propagate the state until the place where it ultimately
gets stored.

Also fixes problem where Encryption was ignoring the state
where the device was already encrypted.

Fixes bug 17881324

Change-Id: Iec09e4464832a506bb2a78bb14a38b3531971fa0