[FRP] Notification screen slides in wrong direction
The method that launches notification screen from choose lock activity
finished the original activity before starting the new activity. This
made the system think it is a backward transition instead of a forward
one. Swapping the order to start new activity first and then finish
old activity fixed the problem.

This also changes the behavior in settings. However, it seems like this
is also the desired behavior for settings.

Bug: 18723199
Change-Id: I90538fa52e0d62a2274c8e0333682035849802c6